+++ kobo/apps/hook/views/__init__.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import
 
-from .hook import HookViewSet
-from .hook_log import HookLogViewSet
+from .v1.hook import HookViewSet
+from .v1.hook_log import HookLogViewSet
