+++ kobo/apps/hook/views/v1/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+from __future__ import absolute_import
+
+from .hook import HookViewSet
+from .hook_log import HookLogViewSet
