+++ kpi/urls/router_api_v2.py
@@ -6,13 +6,13 @@ from rest_framework_extensions.routers import ExtendedDefaultRouter
 from kobo.apps.hook.views import HookViewSet, HookLogViewSet
 from kpi.views.v1 import (
     AssetFileViewSet,
-    SubmissionViewSet,
 )
 
 from kpi.views.v2.asset import AssetViewSet
 from kpi.views.v2.asset_snapshot import AssetSnapshotViewSet
 from kpi.views.v2.asset_version import AssetVersionViewSet
 from kpi.views.v2.collection import CollectionViewSet
+from kpi.views.v2.data import DataViewSet
 from kpi.views.v2.hook_signal import HookSignalViewSet
 from kpi.views.v2.user import UserViewSet
 
@@ -30,7 +30,7 @@ asset_routes.register(r'hook-signal',
                       parents_query_lookups=['asset'],
                       )
 asset_routes.register(r'data',
-                      SubmissionViewSet,
+                      DataViewSet,
                       base_name='submission',
                       parents_query_lookups=['asset'],
                       )
